+++ NEWS.md
@@ -1,3 +1,16 @@
+# scan 0.61.999
+
+## New
+
+## Changes
+
+- `tau_u()`: New method `"tarlow"` calculates Tau-U as implemented in an R code and online calculator by Tarlow (2017). Here, tau values are calculated as in the `method = "complete", continuity_correction = TRUE, tau_method = "a"`. Inferential statistics are calculated based an tau b and the standard deviation for S is derived directly from Kendall's Tau B analysis (different from the `parker` and `complete` methods). 
+- `tau_u()`: Method `"parker"` ignores the `tau_method` setting and sets `continuity_correction = FALSE`. This follows the Parker (2011) paper. There, the inferential statistics are calculated using Kendall's Tau b while the actual Tau calculation applies Kendall's Tau a (without ties).
+
+
+
+## Solved bugs
+
 # scan 0.61.0
 
 ## Solved bugs
@@ -6,9 +19,9 @@
 
 ## New
 
-- Tip-of-the-day like message at startup.
-- Mulitple improvements of the Shiny app (try out with `shinyscan()`)
-- new output engine for rendering html export based on gtable. Set `options(scan.export.engine = "gt")`. This engine allows to export tables into docx format: `overlap(exampleAB) |> export(file = "test.docx", flip = TRUE)`.
+- Tip-of-the-day like message at start-up.
+- Multiple improvements of the Shiny app (try out with `shinyscan()`)
+- new output engine for rendering html export based on *gt table*. Set `options(scan.export.engine = "gt")`. This engine allows to export tables into docx format: `overlap(exampleAB) |> export(file = "test.docx", flip = TRUE)`.
 - new export functions for `pem()`, `pet()`, `pnd()`, and `summary()` (either `summary(exampleAB) |> export()` or `export(exampleAB, summary = TRUE)`)
 
 ## Changes
@@ -247,7 +260,7 @@ Toni:
 
 - scdf files now allow to combine studies with different phase designs.
 Several functions have been adapted to handle cases with differing designs in a mutual analysis.
-- The %>% operator has been imported and exported from the magrittr package. Now that R 4.1 has a
+- The `%>%` operator has been imported and exported from the magrittr package. Now that R 4.1 has a
 pipe operator, pipes seem to become the standard. For compatibility with older R Versions, we will stay with the `%>%` operator for some time before switching to `|>`.
 - To allow for a piping code, we added several functions: `add_l2, select_phases, select_cases, subset, set_vars, set_dvar, set_mvar, set_pvar`.